Move lambda-alarms from patterns into constructs
